Extract other scale property rules to `src/compile/scale/rules`
