+++ test/compile/scale.test.ts
@@ -2,12 +2,12 @@
 
 import {assert} from 'chai';
 
-import {rangeStep, parseScaleComponent, initScale, defaultProperty} from '../../src/compile/scale';
+import {rangeStep, parseScaleComponent, initScale} from '../../src/compile/scale';
 import {parseUnitModel} from '../util';
 
 import * as log from '../../src/log';
-import {X, Channel, NONSPATIAL_SCALE_CHANNELS} from '../../src/channel';
-import {ScaleType, defaultScaleConfig} from '../../src/scale';
+import {X} from '../../src/channel';
+import {defaultScaleConfig} from '../../src/scale';
 import {POINT, TEXT} from '../../src/mark';
 
 describe('Scale', function() {
@@ -30,110 +30,6 @@ describe('Scale', function() {
       assert.equal(scale.paddingInner, 0.3);
       assert.equal(scale.paddingOuter, 0.15);
       assert.isUndefined(scale.padding);
-    });
-  });
-
-  describe('defaultProperty', () => {
-    describe('nice', () => {
-      // TODO:
-    });
-
-    describe('padding', () => {
-      it('should be pointPadding for point scale if channel is x or y and padding is not specified.', () => {
-        for (let c of ['x', 'y'] as Channel[]) {
-          assert.equal(defaultProperty.padding(c, 'point', {pointPadding: 13}), 13);
-        }
-      });
-    });
-
-    describe('paddingInner', () => {
-      it('should be undefined if padding is specified.', () => {
-        assert.equal(defaultProperty.paddingInner(10, 'x', {}), undefined);
-      });
-
-      it('should be bandPaddingInner if channel is x or y and padding is not specified.', () => {
-        assert.equal(defaultProperty.paddingInner(undefined, 'x', {bandPaddingInner: 15}), 15);
-        assert.equal(defaultProperty.paddingInner(undefined, 'y', {bandPaddingInner: 15}), 15);
-      });
-
-      it('should be undefined for non-xy channels.', () => {
-        for (let c of NONSPATIAL_SCALE_CHANNELS) {
-          assert.equal(defaultProperty.paddingInner(undefined, c, {bandPaddingInner: 15}), undefined);
-        }
-      });
-    });
-
-    describe('paddingOuter', () => {
-      it('should be undefined if padding is specified.', () => {
-        for (let scaleType of ['point', 'band'] as ScaleType[]) {
-          assert.equal(defaultProperty.paddingOuter(10, 'x', scaleType, 0, {}), undefined);
-        }
-      });
-
-      it('should be config.scale.bandPaddingOuter for band scale if channel is x or y and padding is not specified and config.scale.bandPaddingOuter.', () => {
-        for (let c of ['x', 'y'] as Channel[]) {
-          assert.equal(defaultProperty.paddingOuter(undefined, c, 'band', 0, {bandPaddingOuter: 16}), 16);
-        }
-      });
-      it('should be paddingInner/2 for band scale if channel is x or y and padding is not specified and config.scale.bandPaddingOuter.', () => {
-        for (let c of ['x', 'y'] as Channel[]) {
-          assert.equal(defaultProperty.paddingOuter(undefined, c, 'band', 10, {}), 5);
-        }
-      });
-
-      it('should be undefined for non-xy channels.', () => {
-        for (let c of NONSPATIAL_SCALE_CHANNELS) {
-          for (let scaleType of ['point', 'band'] as ScaleType[]) {
-            assert.equal(defaultProperty.paddingOuter(undefined, c, scaleType, 0, {}), undefined);
-          }
-        }
-      });
-    });
-
-    describe('round', () => {
-      it('should return scaleConfig.round for x, y, row, column.', () => {
-        for (let c of ['x', 'y', 'row', 'column'] as Channel[]) {
-          assert(defaultProperty.round(c, {round: true}));
-          assert(!defaultProperty.round(c, {round: false}));
-        }
-      });
-
-      it('should return undefined other channels (not x, y, row, column).', () => {
-        for (let c of NONSPATIAL_SCALE_CHANNELS) {
-          assert.isUndefined(defaultProperty.round(c, {round: true}));
-          assert.isUndefined(defaultProperty.round(c, {round: false}));
-        }
-      });
-    });
-
-    describe('zero', () => {
-      it('should return true when mapping a quantitative field to size', () => {
-        assert(defaultProperty.zero({}, 'size', {field: 'a', type: 'quantitative'}));
-      });
-
-      it('should return false when mapping a ordinal field to size', () => {
-        assert(!defaultProperty.zero({}, 'size', {field: 'a', type: 'ordinal'}));
-      });
-
-      it('should return true when mapping a non-binned quantitative field to x/y', () => {
-        for (let channel of ['x', 'y'] as Channel[]) {
-          assert(defaultProperty.zero({}, channel, {field: 'a', type: 'quantitative'}));
-        }
-      });
-
-      it('should return false when mapping a binned quantitative field to x/y', () => {
-        for (let channel of ['x', 'y'] as Channel[]) {
-          assert(!defaultProperty.zero({}, channel, {bin: true, field: 'a', type: 'quantitative'}));
-        }
-      });
-
-      it('should return false when mapping a non-binned quantitative field with custom domain to x/y', () => {
-        for (let channel of ['x', 'y'] as Channel[]) {
-          assert(!defaultProperty.zero({domain: [1, 5]}, channel, {
-            bin: true, field: 'a', type: 'quantitative'
-          }));
-        }
-      });
     });
   });
